+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/fir/checkers/DependencyGraphCreatorChecker.kt
@@ -35,7 +35,8 @@ internal object DependencyGraphCreatorChecker : FirClassChecker(MppCheckerKind.C
 
   private val NON_INCLUDES_KINDS = setOf(ClassKind.ENUM_CLASS, ClassKind.ANNOTATION_CLASS)
 
-  override fun check(declaration: FirClass, context: CheckerContext, reporter: DiagnosticReporter) {
+  context(context: CheckerContext, reporter: DiagnosticReporter)
+  override fun check(declaration: FirClass) {
     declaration.source ?: return
     val session = context.session
     val classIds = session.classIds
@@ -54,15 +55,12 @@ internal object DependencyGraphCreatorChecker : FirClassChecker(MppCheckerKind.C
           declaration.source,
           FirMetroErrors.GRAPH_CREATORS_ERROR,
           "${annotationClassId.relativeClassName.asString()} declarations can only be interfaces.",
-          context,
         )
         return
       }
     }
 
     declaration.validateApiDeclaration(
-      context,
-      reporter,
       "${annotationClassId.relativeClassName.asString()} declarations",
       checkConstructor = true,
     ) {
@@ -72,7 +70,6 @@ internal object DependencyGraphCreatorChecker : FirClassChecker(MppCheckerKind.C
     val createFunction =
       declaration.singleAbstractFunction(
         session,
-        context,
         reporter,
         "@${annotationClassId.relativeClassName.asString()} declarations",
       ) {
@@ -82,7 +79,7 @@ internal object DependencyGraphCreatorChecker : FirClassChecker(MppCheckerKind.C
     val targetGraph = createFunction.resolvedReturnType.toClassSymbol(session)
     val targetGraphAnnotation =
       targetGraph
-        ?.annotations
+        ?.resolvedCompilerAnnotationsWithClassIds
         ?.annotationsIn(session, classIds.graphLikeAnnotations)
         ?.singleOrNull()
     targetGraph?.let {
@@ -91,7 +88,6 @@ internal object DependencyGraphCreatorChecker : FirClassChecker(MppCheckerKind.C
           createFunction.resolvedReturnTypeRef.source ?: declaration.source,
           FirMetroErrors.GRAPH_CREATORS_ERROR,
           "${annotationClassId.relativeClassName.asString()} abstract function '${createFunction.name}' must return a dependency graph but found ${it.classId.asSingleFqName()}.",
-          context,
         )
         return
       }
@@ -106,7 +102,6 @@ internal object DependencyGraphCreatorChecker : FirClassChecker(MppCheckerKind.C
             targetGraphAnnotation.source ?: declaration.source,
             FirMetroErrors.GRAPH_CREATORS_ERROR,
             "${annotationClassId.relativeClassName.asString()} abstract function '${createFunction.name}' must return a contributed graph extension but found ${it.classId.asSingleFqName()}.",
-            context,
           )
           return
         }
@@ -116,7 +111,6 @@ internal object DependencyGraphCreatorChecker : FirClassChecker(MppCheckerKind.C
             targetGraphAnnotation.source ?: declaration.source,
             FirMetroErrors.GRAPH_CREATORS_ERROR,
             "${annotationClassId.relativeClassName.asString()} declarations must be nested within the contributed graph they create but was ${declaration.getContainingClassSymbol()?.classId?.asSingleFqName() ?: "top-level"}.",
-            context,
           )
           return
         }
@@ -125,7 +119,11 @@ internal object DependencyGraphCreatorChecker : FirClassChecker(MppCheckerKind.C
 
     val targetGraphScopes = targetGraphAnnotation?.allScopeClassIds().orEmpty()
     val targetGraphScopeAnnotations =
-      targetGraph?.annotations?.scopeAnnotations(session).orEmpty().toSet()
+      targetGraph
+        ?.resolvedCompilerAnnotationsWithClassIds
+        ?.scopeAnnotations(session)
+        .orEmpty()
+        .toSet()
 
     if (isContributed) {
       val contributedScopes = graphFactoryAnnotation.allScopeClassIds()
@@ -136,7 +134,6 @@ internal object DependencyGraphCreatorChecker : FirClassChecker(MppCheckerKind.C
           graphFactoryAnnotation.source ?: declaration.source,
           FirMetroErrors.GRAPH_CREATORS_ERROR,
           "${annotationClassId.relativeClassName.asString()} declarations must contribute to a different scope than their contributed graph. However, this factory and its contributed graph both contribute to '${overlapping.map { it.asFqNameString() }.single()}'.",
-          context,
         )
         return
       }
@@ -153,7 +150,6 @@ internal object DependencyGraphCreatorChecker : FirClassChecker(MppCheckerKind.C
           param.source,
           FirMetroErrors.GRAPH_CREATORS_ERROR,
           "${annotationClassId.relativeClassName.asString()} abstract function parameters must be unique.",
-          context,
         )
         continue
       }
@@ -161,7 +157,7 @@ internal object DependencyGraphCreatorChecker : FirClassChecker(MppCheckerKind.C
       var isIncludes = false
       var isProvides = false
       var isExtends = false
-      for (annotation in param.annotations) {
+      for (annotation in param.resolvedCompilerAnnotationsWithClassIds) {
         if (!annotation.isResolved) continue
         val annotationClassId = annotation.toAnnotationClassIdSafe(session) ?: continue
         when (annotationClassId) {
@@ -184,7 +180,6 @@ internal object DependencyGraphCreatorChecker : FirClassChecker(MppCheckerKind.C
           param.source,
           FirMetroErrors.GRAPH_CREATORS_ERROR,
           "${annotationClassId.relativeClassName.asString()} abstract function parameters must be annotated with exactly one @Includes, @Provides, or @Extends.",
-          context,
         )
       }
       if ((isIncludes && isProvides) || (isIncludes && isExtends) || (isProvides && isExtends)) {
@@ -200,7 +195,6 @@ internal object DependencyGraphCreatorChecker : FirClassChecker(MppCheckerKind.C
           param.resolvedReturnTypeRef.source ?: param.source ?: declaration.source,
           FirMetroErrors.GRAPH_CREATORS_ERROR,
           "${annotationClassId.relativeClassName.asString()} declarations cannot have their target graph type as parameters.",
-          context,
         )
         continue
       }
@@ -212,7 +206,6 @@ internal object DependencyGraphCreatorChecker : FirClassChecker(MppCheckerKind.C
               param.source,
               FirMetroErrors.GRAPH_CREATORS_ERROR,
               "@Includes cannot be applied to enums, annotations, or platform types.",
-              context,
             )
           }
         }
@@ -221,7 +214,7 @@ internal object DependencyGraphCreatorChecker : FirClassChecker(MppCheckerKind.C
         }
         isExtends -> {
           val dependencyGraphAnno =
-            type.annotations
+            type.resolvedCompilerAnnotationsWithClassIds
               .annotationsIn(session, classIds.dependencyGraphAnnotations)
               .firstOrNull()
           when {
@@ -230,7 +223,6 @@ internal object DependencyGraphCreatorChecker : FirClassChecker(MppCheckerKind.C
                 param.source,
                 FirMetroErrors.GRAPH_CREATORS_ERROR,
                 "@Extends types must be annotated with @DependencyGraph.",
-                context,
               )
             }
 
@@ -239,7 +231,6 @@ internal object DependencyGraphCreatorChecker : FirClassChecker(MppCheckerKind.C
                 param.source,
                 FirMetroErrors.GRAPH_CREATORS_ERROR,
                 "@Extends graphs must be extendable (set DependencyGraph.isExtendable to true).",
-                context,
               )
             }
 
@@ -259,7 +250,6 @@ internal object DependencyGraphCreatorChecker : FirClassChecker(MppCheckerKind.C
                       appendLine(overlap.asSingleFqName().asString())
                     }
                   },
-                  context,
                 )
               }
               for (parentScope in parentScopes) {
@@ -279,7 +269,6 @@ internal object DependencyGraphCreatorChecker : FirClassChecker(MppCheckerKind.C
                       append("Parent 2: ")
                       appendLine(parentClassId.asSingleFqName())
                     },
-                    context,
                   )
                   return
                 }
@@ -287,7 +276,8 @@ internal object DependencyGraphCreatorChecker : FirClassChecker(MppCheckerKind.C
             }
 
             targetGraphScopeAnnotations.isNotEmpty() -> {
-              val parentScopeAnnotations = type.annotations.scopeAnnotations(session).toSet()
+              val parentScopeAnnotations =
+                type.resolvedCompilerAnnotationsWithClassIds.scopeAnnotations(session).toSet()
               val overlaps = parentScopeAnnotations.intersect(targetGraphScopeAnnotations)
               if (overlaps.isNotEmpty()) {
                 reporter.reportOn(
@@ -302,7 +292,6 @@ internal object DependencyGraphCreatorChecker : FirClassChecker(MppCheckerKind.C
                       appendLine(overlap.simpleString())
                     }
                   },
-                  context,
                 )
                 return
               }
@@ -325,7 +314,6 @@ internal object DependencyGraphCreatorChecker : FirClassChecker(MppCheckerKind.C
                       append("Parent 2: ")
                       appendLine(parentClassId.asSingleFqName())
                     },
-                    context,
                   )
                 }
               }