+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/fir/checkers/InjectConstructorChecker.kt
@@ -17,7 +17,9 @@ import org.jetbrains.kotlin.fir.declarations.FirClass
 import org.jetbrains.kotlin.fir.declarations.getAnnotationByClassId
 
 internal object InjectConstructorChecker : FirClassChecker(MppCheckerKind.Common) {
-  override fun check(declaration: FirClass, context: CheckerContext, reporter: DiagnosticReporter) {
+
+  context(context: CheckerContext, reporter: DiagnosticReporter)
+  override fun check(declaration: FirClass) {
     val source = declaration.source ?: return
     val session = context.session
     val classIds = session.classIds
@@ -26,7 +28,7 @@ internal object InjectConstructorChecker : FirClassChecker(MppCheckerKind.Common
       declaration.annotationsIn(session, classIds.injectAnnotations).toList()
 
     val injectedConstructor =
-      declaration.symbol.findInjectConstructor(session, context, reporter, checkClass = false) {
+      declaration.symbol.findInjectConstructor(session, checkClass = false) {
         return
       }
 
@@ -36,7 +38,7 @@ internal object InjectConstructorChecker : FirClassChecker(MppCheckerKind.Common
     declaration
       .getAnnotationByClassId(DaggerSymbols.ClassIds.DAGGER_REUSABLE_CLASS_ID, session)
       ?.let {
-        reporter.reportOn(it.source ?: source, FirMetroErrors.DAGGER_REUSABLE_ERROR, context)
+        reporter.reportOn(it.source ?: source, FirMetroErrors.DAGGER_REUSABLE_ERROR)
         return
       }
 
@@ -44,7 +46,6 @@ internal object InjectConstructorChecker : FirClassChecker(MppCheckerKind.Common
       reporter.reportOn(
         injectedConstructor.source,
         FirMetroErrors.CANNOT_HAVE_INJECT_IN_MULTIPLE_TARGETS,
-        context,
       )
       return
     }