+++ CHANGELOG.md
@@ -4,6 +4,10 @@ Changelog
 **Unreleased**
 --------------
 
+- **Enhancement:** Check for context parameters in top-level function injection checker.
+- Update to Kotlin `2.2.0`.
+- Update Gradle plugin to target Kotlin language version to `2.0`.
+
 0.3.5
 -----